@@ -13,26 +13,26 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "18.2.0",
-    "@angular/common": "18.2.0",
-    "@angular/compiler": "18.2.0",
-    "@angular/core": "18.2.0",
-    "@angular/forms": "18.2.0",
-    "@angular/platform-browser": "18.2.0",
-    "@angular/platform-browser-dynamic": "18.2.0",
-    "@angular/platform-server": "18.2.0",
-    "@angular/router": "18.2.0",
-    "@angular/ssr": "^18.2.0",
+    "@angular/animations": "19.2.3",
+    "@angular/common": "19.2.3",
+    "@angular/compiler": "19.2.3",
+    "@angular/core": "19.2.3",
+    "@angular/forms": "19.2.3",
+    "@angular/platform-browser": "19.2.3",
+    "@angular/platform-browser-dynamic": "19.2.3",
+    "@angular/platform-server": "19.2.3",
+    "@angular/router": "19.2.3",
+    "@angular/ssr": "^19.2.5",
     "date-fns": "2.30.0",
     "rxjs": "7.8.0",
     "tslib": "2.3.0",
     "wavesurfer.js": "7.3.2",
-    "zone.js": "0.14.10"
+    "zone.js": "0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "18.2.0",
-    "@angular/cli": "18.2.0",
-    "@angular/compiler-cli": "18.2.0",
+    "@angular-devkit/build-angular": "19.2.5",
+    "@angular/cli": "19.2.5",
+    "@angular/compiler-cli": "19.2.3",
     "@cloudflare/workers-types": "^4.20240502.0",
     "@types/jasmine": "4.3.0",
     "@types/node": "^18.18.0",
