@@ -3,10 +3,9 @@ import { Component } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
 
 @Component({
-  selector: 'app-root',
-  standalone: true,
-  imports: [RouterOutlet],
-  template: '<router-outlet />',
+    selector: 'app-root',
+    imports: [RouterOutlet],
+    template: '<router-outlet />'
 })
 export class AppComponent {
   title = 'store';
