@@ -1,4 +1,4 @@
-import type { Routes } from '@angular/router';
+import { Routes } from '@angular/router';
 import { LayoutComponent } from '@shared/components/layout/layout.component';
 import { NotFoundComponent } from '@shared/pages/not-found/not-found.component';
 
@@ -13,17 +13,15 @@ export const routes: Routes = [
           import('./domains/products/pages/list/list.component'),
       },
       {
-        path: 'product/:id',
+        path: 'category/:slug',
+        loadComponent: () =>
+          import('./domains/products/pages/list/list.component'),
+      },
+      {
+        path: 'product/:slug',
         loadComponent: () =>
           import(
             './domains/products/pages/product-detail/product-detail.component'
-          ),
-      },
-      {
-        path: 'locations',
-        loadComponent: () =>
-          import(
-            './domains/locations/locations.component'
           ),
       },
     ],
