@@ -1,11 +1,10 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-locations',
-  standalone: true,
-  imports: [],
-  templateUrl: './locations.component.html',
-  styleUrl: './locations.component.css'
+    selector: 'app-locations',
+    imports: [],
+    templateUrl: './locations.component.html',
+    styleUrl: './locations.component.css'
 })
 export default class LocationsComponent {
 
