@@ -1,21 +1,29 @@
 import { Component, inject, signal } from '@angular/core';
-
-import { CartService } from '@services/cart.service';
+import { CommonModule } from '@angular/common';
+import { CartService } from '@shared/services/cart.service';
 import { RouterLinkWithHref, RouterLinkActive } from '@angular/router';
 
 @Component({
   selector: 'app-header',
-  standalone: true,
-  imports: [RouterLinkWithHref, RouterLinkActive],
+  imports: [
+    CommonModule,
+    RouterLinkWithHref,
+    RouterLinkActive,
+  ],
   templateUrl: './header.component.html',
 })
 export class HeaderComponent {
   hideSideMenu = signal(true);
+  showMenu = signal(false);
   private cartService = inject(CartService);
   cart = this.cartService.cart;
   total = this.cartService.total;
 
   toogleSideMenu() {
-    this.hideSideMenu.update((prevState) => !prevState);
+    this.hideSideMenu.update(prevState => !prevState);
+  }
+
+  toggleMenu() {
+    this.showMenu.update(prevState => !prevState);
   }
 }
