@@ -1,11 +1,11 @@
 import { Component } from '@angular/core';
+import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
 import { HeaderComponent } from '@shared/components/header/header.component';
 
 @Component({
   selector: 'app-layout',
-  standalone: true,
-  imports: [HeaderComponent, RouterModule],
+  imports: [CommonModule, HeaderComponent, RouterModule],
   templateUrl: './layout.component.html',
 })
 export class LayoutComponent {}
