@@ -3,10 +3,8 @@ import { Component } from '@angular/core';
 import { RouterLinkWithHref } from '@angular/router';
 
 @Component({
-  selector: 'app-not-found',
-  standalone: true,
-  imports: [RouterLinkWithHref],
-  templateUrl: './not-found.component.html',
-  styleUrls: ['./not-found.component.css'],
+    selector: 'app-not-found',
+    imports: [RouterLinkWithHref],
+    templateUrl: './not-found.component.html'
 })
 export class NotFoundComponent {}
