@@ -1,7 +1,7 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable, inject } from '@angular/core';
-import { Category } from '@models/category.model';
-import { environment } from '@envs/environment';
+import { Category } from '@shared/models/category.model';
+import { environment } from '@env/environment';
 
 @Injectable({
   providedIn: 'root',
@@ -9,11 +9,7 @@ import { environment } from '@envs/environment';
 export class CategoryService {
   private http = inject(HttpClient);
 
-  constructor() {}
-
   getAll() {
-    return this.http.get<Category[]>(
-      `${environment.API_URL}/api/v1/categories`,
-    );
+    return this.http.get<Category[]>(`${environment.apiUrl}/api/v1/categories`);
   }
 }
